refactor(metrics): address native metrics PR review feedback
